upgrade to v0.40

+++ src/order/MoveToPoint.java
@@ -3,7 +3,7 @@ public class MoveToPoint extends StationaryOrder{
 	
 	final MoveTo moveOrder;
 	
-	public MoveToPoint(int posX, int posY, double velX, double velY){
+	public MoveToPoint(double posX, double posY, double velX, double velY){
 		super(posX, posY, velX, velY, 250, 0.062);
 		moveOrder = new MoveTo(this);
 	}
